Switch to rendering the TUI on stderr
Doing this to allow use in the shell, e.g.

```zsh
for dir in $(picleo <DIRS>)
do
    pushd "$dir"
done
```

When rendering to stdout it appears to hang as all characters output are caught by the shell. By moving this to stderr we can capture stdout and still interact with the user on stderr.